# project/review-o/review-o/openapi.yaml
# Generated with protoc-gen-openapi
# https://github.com/google/gnostic/tree/master/cmd/protoc-gen-openapi

openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
    /v1/appeal/audit:
        post:
            tags:
                - Review
            description: o 端评价申诉审核
            operationId: Review_AuditAppeal
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.review.v1.AuditAppealRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.review.v1.AuditAppealReply'
    /v1/review:
        post:
            tags:
                - Review
            description: c端创建评价
            operationId: Review_CreateReview
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.review.v1.CreateReviewRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.review.v1.CreateReviewReply'
    /v1/review/appeal:
        post:
            tags:
                - Review
            description: b端申述评价
            operationId: Review_AppealReview
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.review.v1.AppealReviewRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.review.v1.AppealReviewReply'
    /v1/review/reply:
        post:
            tags:
                - Review
            description: b端回复评价
            operationId: Review_ReplyReview
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.review.v1.ReplyReviewRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.review.v1.ReplyReviewReply'
    /v1/review/{reviewID}:
        get:
            tags:
                - Review
            description: c端获取评价详情
            operationId: Review_GetReview
            parameters:
                - name: reviewID
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.review.v1.GetReviewReply'
    /v1/{userID}/reviews:
        get:
            tags:
                - Review
            description: C端查看userID下所有评价
            operationId: Review_ListReviewByUserId
            parameters:
                - name: userID
                  in: path
                  required: true
                  schema:
                    type: string
                - name: page
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: size
                  in: query
                  schema:
                    type: integer
                    format: int32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.review.v1.ListReviewByUserIdReply'
    b/v1/review/reply:
        post:
            tags:
                - Business
            description: 回复评价
            operationId: Business_ReplyReview
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.busniess.v1.ReplyReviewRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.busniess.v1.ReplyReviewReply'
    o/v1/appeal/audit:
        post:
            tags:
                - Operation
            operationId: Operation_AuditAppeal
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.operation.v1.AuditAppealReques'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.operation.v1.AuditAppealReply'
    o/v1/review/audit:
        post:
            tags:
                - Operation
            operationId: Operation_AuditReview
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.operation.v1.AuditReviewRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.operation.v1.AuditReviewReply'
    v1/review/audit:
        post:
            tags:
                - Review
            description: o端审核评价
            operationId: Review_AuditRrview
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.review.v1.AuditRrviewRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.review.v1.AuditRrviewReply'
components:
    schemas:
        api.busniess.v1.ReplyReviewReply:
            type: object
            properties:
                replyID:
                    type: string
        api.busniess.v1.ReplyReviewRequest:
            type: object
            properties:
                reviewID:
                    type: string
                storeID:
                    type: string
                content:
                    type: string
                picInfo:
                    type: string
                videoInfo:
                    type: string
        api.operation.v1.AuditAppealReply:
            type: object
            properties: {}
        api.operation.v1.AuditAppealReques:
            type: object
            properties:
                appealID:
                    type: string
                reviewID:
                    type: string
                status:
                    type: integer
                    format: int32
                opUser:
                    type: string
                opReason:
                    type: string
                opRemarks:
                    type: string
        api.operation.v1.AuditReviewReply:
            type: object
            properties: {}
        api.operation.v1.AuditReviewRequest:
            type: object
            properties:
                reviewID:
                    type: string
                status:
                    type: integer
                    format: int32
                opUser:
                    type: string
                opReason:
                    type: string
                opRemarks:
                    type: string
        api.review.v1.AppealReviewReply:
            type: object
            properties:
                appleID:
                    type: string
        api.review.v1.AppealReviewRequest:
            type: object
            properties:
                reviewID:
                    type: string
                storeID:
                    type: string
                reason:
                    type: string
                content:
                    type: string
                picinfo:
                    type: string
                videoInfo:
                    type: string
        api.review.v1.AuditAppealReply:
            type: object
            properties: {}
        api.review.v1.AuditAppealRequest:
            type: object
            properties:
                appealID:
                    type: string
                reviewID:
                    type: string
                status:
                    type: integer
                    format: int32
                opUser:
                    type: string
                opRemarks:
                    type: string
        api.review.v1.AuditRrviewReply:
            type: object
            properties:
                reviewID:
                    type: string
                status:
                    type: integer
                    format: int32
        api.review.v1.AuditRrviewRequest:
            type: object
            properties:
                reviewID:
                    type: string
                status:
                    type: integer
                    format: int32
                opUser:
                    type: string
                opReason:
                    type: string
                opMarks:
                    type: string
        api.review.v1.CreateReviewReply:
            type: object
            properties:
                reviewId:
                    type: string
        api.review.v1.CreateReviewRequest:
            type: object
            properties:
                userId:
                    type: string
                orderId:
                    type: string
                score:
                    type: integer
                    format: int32
                serviceScore:
                    type: integer
                    format: int32
                expreeScore:
                    type: integer
                    format: int32
                content:
                    type: string
                picinfo:
                    type: string
                videoInfo:
                    type: string
                annoymous:
                    type: boolean
        api.review.v1.GetReviewReply:
            type: object
            properties:
                data:
                    $ref: '#/components/schemas/api.review.v1.ReviewInfo'
        api.review.v1.ListReviewByUserIdReply:
            type: object
            properties:
                list:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.review.v1.ReviewInfo'
        api.review.v1.ReplyReviewReply:
            type: object
            properties:
                replyID:
                    type: string
        api.review.v1.ReplyReviewRequest:
            type: object
            properties:
                reviewID:
                    type: string
                storeID:
                    type: string
                content:
                    type: string
                picinfo:
                    type: string
                videoInfo:
                    type: string
        api.review.v1.ReviewInfo:
            type: object
            properties:
                reviewID:
                    type: string
                userID:
                    type: string
                orderID:
                    type: string
                score:
                    type: integer
                    format: int32
                serviceScore:
                    type: integer
                    format: int32
                expressScore:
                    type: integer
                    format: int32
                content:
                    type: string
                picInfo:
                    type: string
                videoInfo:
                    type: string
                status:
                    type: integer
                    format: int32
tags:
    - name: Business
      description: 商家端服务
    - name: Operation
    - name: Review